move pillar to holes as well
